feat(telemetry): new @tao.js/telemetry package; move TaoLogger from utils
Renames @tao.js/trace to @tao.js/telemetry (packages/tao-telemetry) as the
home for signal-network telemetry: the causal Tracer (pure Network
decoration), InMemorySink/ConsoleSink, W3C trace context helpers - and
TaoLogger, moved from @tao.js/utils with its full test suite. utils keeps
a deprecated re-export (with a file:-protocol runtime dependency on
telemetry; workspace: protocol breaks pnpm single-instance core sharing
in-workspace - see AGENTS.md) so existing TaoLogger imports keep working
through an overlap window.

Tracer x Channel/Transponder composability tests move into the utils suite
to keep the package graph acyclic (utils -> telemetry -> core). Also adds
mutation-survivor tests for core (decoration validation edges, double
dispose, wildcard entry/chain policy on kernels and bare Networks, hooks
absence observation, malformed hooks fallback, legacy caller-forward
invocation, AppCtx action-key datum) and annotates three try-guarded
equivalent mutants. All 12 project suites pass; the socket.io end-to-end
smoke passes 7/7 (core instance resolution in the smoke now mirrors the
adapters' own resolution).

Affected packages:
- @tao.js/core
- @tao.js/utils
- @tao.js/telemetry
- @tao.js/opentelemetry
