+++ AGENTS.md
@@ -190,8 +190,8 @@ For isolated tests, prefer `new Kernel()` over the shared default `TAO`.
 | `@tao.js/routing-react-router`    | `packages/tao-routing-react-router` | React Router adapter (`importLoader`, `useLoaderSignal`)                                                                                                                     |
 | `@tao.js/routing-tanstack-router` | `packages/tao-routing-tanstack`     | TanStack Router adapter (`importLoader`, `useLoaderSignal`)                                                                                                                  |
 | `@tao.js/routing-next`            | `packages/tao-routing-next`         | Next.js adapter (`importLoader`, `enterRoute`, `useRouteSignal`)                                                                                                             |
-| `@tao.js/trace`                   | `packages/tao-trace`                | Causal signal tracer as a pure Network decoration (chain reducer + observer); W3C traceparent helpers; `InMemorySink`/`ConsoleSink`                          |
-| `@tao.js/opentelemetry`          | `packages/tao-opentelemetry`        | OpenTelemetry exporter sink for `@tao.js/trace` records (spans with causal parentage; api-only dependency)                                                   |
+| `@tao.js/telemetry`               | `packages/tao-telemetry`            | Telemetry: causal `Tracer` (pure Network decoration), `TaoLogger` (moved from utils; deprecated re-export remains), sinks, W3C traceparent helpers          |
+| `@tao.js/opentelemetry`          | `packages/tao-opentelemetry`        | OpenTelemetry exporter sink for `@tao.js/telemetry` records (spans with causal parentage; api-only dependency)                                                   |
 | `@tao.js/socket.io`               | `packages/tao-socket-io`            | Wire a Kernel to socket.io (`wireTaoJsToSocketIO`)                                                                                                                           |
 | `@tao.js/koa`                     | `packages/koa-tao`                  | Expose a TAO network over HTTP via Koa middleware                                                                                                                            |
 | `@tao.js/http-client`             | `packages/tao-http-client`          | HTTP client for a TAO HTTP server (`TaoHttpClient`) — private                                                                                                                |
@@ -319,7 +319,7 @@ _Append learnings for the next agent. Newest first._
 
 - **2026-07-19** — `@tao.js/react`: prefer named export `TaoProvider`; `Provider` remains a deprecated alias (dev once-warning via `deprecations.js`). Default export of `src/Provider.js` is `TaoProvider`.
 - **2026-07-18** — Host-router adapters: `@tao.js/routing-core` (signal apply + `createImportLoader` + React hook factories without importing React) and peer adapters `@tao.js/routing-react-router`, `@tao.js/routing-tanstack-router`, `@tao.js/routing-next`. Prefer these over investing in legacy `@tao.js/router`. Mutation: `pnpm test:mutation:routing-*`.
-- **2026-07-18** — Envelope/decoration redesign implemented on `feat/network-envelope` per ENVELOPE-SPEC.md (spec committed first; production surveys of kettleos + tidy.dev embedded in §10 with eight normative invariants). Core: `Network.enter` hop engine (dispatch-once, cascade/hop/chain scopes), `Network.decorate`, `AppCtxHandlers` settlement hook, legacy `setCtxControl`+forward path frozen. Utils adapters migrated (Channel = cascade + onForward mirror; Source/Relay = hop-scope origin marker, Relay's unbound-forward bug fixed; Transponder = cascade entry, chains now propagate on bare kernels; Transceiver = settlement hook, `captureSignal` fork deleted). Adapters throw a clear error on pre-envelope cores (mixed-version installs are real — tidy.dev runs core 0.16.0 with utils 0.16.2; set the peerDependency floor when versions are cut at release). New `@tao.js/trace` + `@tao.js/opentelemetry`: tracing is a pure decoration — full causal trees for kernel/channel/transponder entries with ZERO instrumentation; one tracer per network (chain key exclusivity). Gotchas learned: white-box utils tests assert threading mechanics, not just semantics — rewrite intent, don't delete; Stryker `inPlace: true` mutates sources during runs (don't git-operate or run other suites on that package concurrently); jest 30 fails tests on unhandled rejections (assert legacy rethrow via the dispatch promise instead); channel-attached handler chains still re-enter with a fresh cascade (frozen Channel semantic — trace shows them as new roots).
+- **2026-07-18** — Envelope/decoration redesign implemented on `feat/network-envelope` per ENVELOPE-SPEC.md (spec committed first; production surveys of kettleos + tidy.dev embedded in §10 with eight normative invariants). Core: `Network.enter` hop engine (dispatch-once, cascade/hop/chain scopes), `Network.decorate`, `AppCtxHandlers` settlement hook, legacy `setCtxControl`+forward path frozen. Utils adapters migrated (Channel = cascade + onForward mirror; Source/Relay = hop-scope origin marker, Relay's unbound-forward bug fixed; Transponder = cascade entry, chains now propagate on bare kernels; Transceiver = settlement hook, `captureSignal` fork deleted). Adapters throw a clear error on pre-envelope cores (mixed-version installs are real — tidy.dev runs core 0.16.0 with utils 0.16.2; set the peerDependency floor when versions are cut at release). New `@tao.js/telemetry` (+ `@tao.js/opentelemetry` exporter): tracing is a pure decoration — full causal trees for kernel/channel/transponder entries with ZERO instrumentation; one tracer per network (chain key exclusivity). `TaoLogger` moved utils → telemetry with a deprecated re-export left in utils (utils gained a runtime dep on telemetry using the `file:` protocol (repo convention). Do NOT use `workspace:` protocol here — pnpm then isolates utils behind a store copy with its own @tao.js/core instance, breaking single-instance `instanceof AppCtx` sharing inside the workspace (found via tools/smoke). Release NOTE: `file:` deps are not rewritten on publish — version this dependency when cutting releases). Gotchas learned: white-box utils tests assert threading mechanics, not just semantics — rewrite intent, don't delete; Stryker `inPlace: true` mutates sources during runs (don't git-operate or run other suites on that package concurrently); jest 30 fails tests on unhandled rejections (assert legacy rethrow via the dispatch promise instead); channel-attached handler chains still re-enter with a fresh cascade (frozen Channel semantic — trace shows them as new roots).
 
 - **2026-07-18** — `@tao.js/react@0.17.0` on `modernize-react-handlers`: DataHandler/Provider/DataConsumer/createContextHandler are function components; tree-scoped `DataLayerContext` + `useTaoData(name?)`; bag merge kept for deprecated `RenderHandler.context` / `DataConsumer`; dev once-warnings via `deprecations.js`. See §5 migration (steps 1–3 shipped; removal still future).
 - **2026-07-18** — `@tao.js/react` `SwitchHandler` / `RenderHandler` modernized to function components (`modernize-react-handlers`). Match table from children each render; Kernel inline subs reconciled via `useTaoInlineSubscription`. Selection by trigram `matchKey` (multi-match / wildcards / array cartesian). Wave = one AppCon (latest wins, not settle-chain union). `shouldRender` stays public; non-RH children always pass through. DataHandler still class — next (migration plan in §5).
