+++ packages/tao-telemetry/src/ConsoleSink.js
