+++ packages/tao-telemetry/src/InMemorySink.js
