+++ packages/tao-telemetry/src/index.js
@@ -1,6 +1,7 @@
 import Tracer, { TRACE_CHAIN } from './Tracer';
 import InMemorySink from './InMemorySink';
 import ConsoleSink from './ConsoleSink';
+import { TaoLogger } from './TaoLogger';
 import {
   newTraceId,
   newSignalId,
@@ -13,6 +14,7 @@ export {
   Tracer,
   InMemorySink,
   ConsoleSink,
+  TaoLogger,
   TRACE_CHAIN,
   newTraceId,
   newSignalId,
