+++ packages/tao-telemetry/test/sinks.spec.js
